+++ packages/client/components/app/interface/settings/user/appearance/AppearanceMenu.tsx
@@ -306,6 +306,7 @@ export function AppearanceMenu() {
               }
               timestamp={new Date()}
               username={user()?.displayName}
+              pronouns={user()?.pronouns}
               isLink="hide"
             >
               Sphinx of black quartz, judge my vow
@@ -440,7 +441,7 @@ const Preview = styled("div", {
     height: "126px",
     overflow: "hidden",
     borderRadius: "var(--borderRadius-lg)",
-    background: "var(--md-sys-color-surface-container-highest)",
+    background: "var(--md-sys-color-surface-container-lowest)",
   },
 });