+++ packages/client/components/app/menus/MessageContextMenu.tsx
@@ -95,6 +95,24 @@ export function MessageContextMenu(props: {
   }
 
   /**
+   * Pin/unpin the message
+   */
+  function pinMessage(ev: MouseEvent) {
+    if (ev.shiftKey) {
+      if (props.message!.pinned) {
+        props.message!.unpin().catch(showError);
+      } else {
+        props.message!.pin().catch(showError);
+      }
+    } else {
+      openModal({
+        type: "pin_message",
+        message: props.message!,
+      });
+    }
+  }
+
+  /**
    * Open message in Stoat Admin Panel
    */
   function openAdminPanel() {
@@ -215,16 +233,7 @@ export function MessageContextMenu(props: {
             props.message!.channel?.havePermission("ManageMessages")
           }
         >
-          <ContextMenuButton
-            icon={MdPin}
-            onClick={() => {
-              if (props.message!.pinned) {
-                props.message!.unpin().catch(showError);
-              } else {
-                props.message!.pin().catch(showError);
-              }
-            }}
-          >
+          <ContextMenuButton icon={MdPin} onClick={pinMessage}>
             <Switch fallback={<Trans>Pin message</Trans>}>
               <Match when={props.message!.pinned}>
                 <Trans>Unpin message</Trans>